requirements.txt
langchain
langchain-community
openai
PyPDF2
faiss-cpu
tqdm
python-dotenv
streamlit
requests

